% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_pro_seg.R
\name{cube_pro_seg}
\alias{cube_pro_seg}
\title{Extract and align all profiles for a given cross-section}
\usage{
cube_pro_seg(cube, dir = "y")
}
\arguments{
\item{cube}{The hyperspectral cube returned by the function \code{cube_seg}. 
See Details.}

\item{dir}{Direction of the profiles. One of "x" or "y". See Details.}
}
\value{
A raster stack with the same wavelength as the input data, with 
profiles aligned at their start position.
}
\description{
Extracts all possible profiles for a cross-section surface, following a 
indicative cluster.
}
\details{
This is a specific function designed to extract and align profiles
of sediment disks cross-section images. The image segmentation should result 
in a single cluster containing the samples. If dir = "y", all columns will 
be sampled from top to bottom and if dir = "x" all rows will be sampled from
left to right. It is important that the image has the proper orientation for 
the results to be meaningful since the profiles will be aligned at their 
start position (top for "y" and left for "x"). The objective is to extract 
depth profiles where depth is relative to the surface at each column (or row, 
if dir = "x").
}
