% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cube.R
\name{cube_read}
\alias{cube_read}
\title{Import data cube}
\usage{
cube_read(file, mask)
}
\arguments{
\item{file}{Path to ENVI hyperspectral cube header file.}

\item{mask}{A raster mask of saturated values such as provided by the 
\code{mask_s} function.}
}
\value{
A raster stack object with additional "metadata" attribute.
}
\description{
This function is a wrapper for the raster \code{stack} function, masking 
saturated pixels and keeping minimal metadata information from the ENVI 
header file.
}
\details{
The data type and wavelength information are retained in the 
"metadata" attribute, and bands are named with their rounded center 
wavelength. Note that the "metadata" attribute will only be automatically 
copied by functions in this library.
}
\seealso{
\code{stack}, \code{mask_s}
}
