% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_refl.R
\name{cube_refl}
\alias{cube_refl}
\title{Calculate lambertian equivalent bi-hemispherical reflectance}
\usage{
cube_refl(cube, dir, fln, rho, method = c("along", "area"), scandir = "x")
}
\arguments{
\item{cube}{Calibrated hyperspectral cube.}

\item{dir}{Path to write reflectance cube. If omitted, data will be saved 
to the working directory.}

\item{fln}{Filename for the reflectance cube. If omitted, data will be 
saved with the same input filename adding the suffix 
".refl.envi".}

\item{rho}{Directional-hemispherical reflectance of reference at the same
wavelength grid of the data.}

\item{method}{Method for reference reflectance. One of "along" or "area". 
See Details.}

\item{scandir}{Scan direction of the image. One of "x" or "y".}
}
\value{
A raster stack with the lambertian equivalent bi-hemispherical 
reflectance (unitless).
}
\description{
An interactive function to calculate the lambertian equivalent bi-
-hemispherical reflectance for imagery that contains at least one target of 
known reflectance.
}
\details{
Although reflectance can be calculated from digital counts, the raw
digital counts of the instrument are not recommended since it may contain 
saturation pixels and spectral or spatial distortions that are corrected in
the calibration process. It is recommended to use only with calibrated data.

Since exact position of samples and reference may vary per image, the 
function is iterative and require the user to point to regions in the image. 
The user is first prompt to select the LL and UR corners to process. To 
process all the image, click the LL below and to the left of the image LL and 
above and to the right to the UR of the image.

There are two methods available for calculating the reflectance. The "along"
method requires that the reflectance reference was aligned with the scan 
direction of the instrument. The exitant radiance of the reference will be 
averaged per scan line. This is specially recommended if data was acquired 
with natural illumination since reflectance normalization will be performed 
per scan line allowing to remove illumination fluctuations during scan. This 
is the preferred method for data acquisition and processing.

The simpler method "area" is also available, requiring to drawn the surface 
of the reference on the image. The signal form the reference surface will be 
averaged and this single value applied to the whole image. This method should
only be used for stable and homogeneous illumination.

The lambertian equivalent bi-hemispherical reflectance is calculated by 
scaling the measured hemispherical-directional reflectance by pi steradians.
}
