% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_saturation.R
\name{mask_s}
\alias{mask_s}
\title{Mask of saturated pixels}
\usage{
mask_s(file, wave, wrange, bits = 12, layers = FALSE, rotate = TRUE)
}
\arguments{
\item{file}{Path to raw (digital counts) ENVI hyperspectral cube.}

\item{wave}{Optional specific wavelengths to be used in the saturation 
mask. It must be provided in the same units as the data 
reference. Default is to use all bands.}

\item{wrange}{Alternative selection of wavelnegths to be used if wave is not 
provided. It must be provided in the same units as the data 
reference. Default is to use all bands.}

\item{bits}{The number of bits of data acquisition. Default is 12 bits from
the SOC-710 model.}

\item{layers}{Logical. Should a single raster be returned or a mask be 
created per layer (stack)?}

\item{rotate}{Logical. SOC imagery in raw and radiance have different 
orientations. Default is set to TRUE to rotate the mask to the
same orientation of the calibrated cube.}
}
\value{
A raster layer or raster stack object with saturated pixels set to 
NA and non-saturated pixels set to 1.
}
\description{
This function creates a raster mask for saturated pixels based on the raw 
digital counts hyperspectral cube.
}
\details{
The function will set the maximum integer possible for the number 
of bits in the data to NA. If \code{layers} is set to FALSE (default) it 
will provide a single raster layer mask where a pixel is set to NA if it is 
saturated at any bands in \code{wave} or \code{wrange}. Note that for the 
function to perform as intended, it is necessary to provide the raw 
(uncalibrated) hyperspectral cube.

The wavelengths in \code{wave} do not have to be exact, the bands with the 
closet nominal wavelength will be returned.
}
