% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample.R
\name{sample_plot}
\alias{sample_plot}
\title{Plot hyperspectral cube samples}
\usage{
sample_plot(cube, sample, id, summary = TRUE, loc = mean, disp = sd, ylab, ...)
}
\arguments{
\item{cube}{The hyperspectral cube returned by the function \code{cube_read}
from which the sample were taken.}

\item{sample}{An object returned be the function \code{cube_sample}.}

\item{id}{Optional specification of the samples to be used. Default to 
use all samples.}

\item{summary}{Logical. Should every pixel be plotted or descriptive 
statistics?}

\item{loc}{If summary is TRUE, the location function to be used. Defaults 
to \code{mean}.}

\item{disp}{If summary is TRUE, the dispersion function to be used. 
Defaults to \code{sd}.}

\item{...}{Arguments to be passed to \code{cube_rgb} (e.g., log = T) and 
descriptive statistics functions (e.g., na.rm = TRUE).}
}
\description{
This function provides a easy interface for simple plots.
}
\details{
This function provides a simple visualization for point (or circle)
or line samples taken with the function \code{cube_sample}. The cube_rgb is 
called to show the spatial position of the samples and a second plot is 
called to plot the samples per wavelength. Note that by default, the ylab 
will be expression(rho~(unitless)) if not provided.

The location and dispersion metric functions should return a single value per 
sample. Read the documentation of \code{sample_stat}.
}
