% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stat.R
\name{sample_stat}
\alias{sample_stat}
\title{Calculate statistics on samples}
\usage{
sample_stat(sample, fun, single = FALSE, ...)
}
\arguments{
\item{sample}{An object returned by the function \code{cube_sample}.}

\item{fun}{A function returning descriptive statistics. See Details.}

\item{single}{Logical. Should samples be combined in a single sample before 
statistics?}

\item{...}{Additional arguments passed to fun.}
}
\value{
A matrix with the statistcs per sample and band (wavelenth).
}
\description{
This function will apply a statistical function to each sample.
}
\details{
The statistical function should return a single value. Examples as 
\code{mean}, \code{var}, \code{sd}, \code{min}, \code{max}.
}
