% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfspec.R
\docType{package}
\name{surfspec}
\alias{surfspec}
\title{surfspec: PAE tools for imaging spectroscopy}
\description{
The surfspec package is a library intended for use by the Protistology and 
Aquatic Ecology (PAE) research group at Ghent University, Belgium. It 
provides generic functions for analysis of hyperspectral imagery such as 
acquired by the SOC-710P camera from the Surface Optics Corporation. It also
provides specific functions for reflectance profile of sediment cores.
}
