% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_seg.R
\name{cube_seg}
\alias{cube_seg}
\title{Image segmentation}
\usage{
cube_seg(cube, type = "kmeans", ...)
}
\arguments{
\item{cube}{The hyperspectral cube returned by the function \code{cube_read}}

\item{type}{Type of image segmentation. 'kmeans' is the only accepted value.}

\item{...}{Arguments to be passed to \code{kmeans}.}
}
\value{
The same cube as the input, with the clusters as a raster layer in 
the attribute 'metadata'.
}
\description{
Performs image segmentation with the kmeans algorithm.
}
