% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bands.R
\name{.find_band}
\alias{.find_band}
\title{Find the band index given wavelength}
\usage{
.find_band(cube, wave)
}
\description{
This is an internal function to find the index of bands most closely matching
the requested wavelength. The matching is made by minimum absolute 
reached by two bands, selecting the lower wavelength one.
}
